+++ packages/components/CHANGELOG.md
@@ -1,7 +1,16 @@
+# v0.2.0
+## Adds keyboard navigation to `Select`
+
+This release adds simple keyboard navigation to the `Select` component.
+
+<!-- -->
+
 # v0.1.1
 ## Fixes exporting
 
 The previous published version did not export anything, as it was missing webpack configuration. This release fixes that and makes this library consumable.
+
+<!-- -->
 
 # v0.1.0
 ## First release